Move stuff out of old Sidebar and merge with new
