+++ react/components/typings.d.ts
@@ -1,0 +1,4 @@
+export type SidebarComponent = {
+  name: string
+  treePath: string
+}
